TruyenQQ: Mirate to 1.6
