+++ src/vi/truyenqq/build.gradle.kts
@@ -8,7 +8,7 @@ keiyoushi {
     name = "TruyenQQ"
     versionCode = 24
     contentWarning = ContentWarning.SAFE
-    libVersion = "1.4"
+    libVersion = "1.6"
 
     source {
         lang = "vi"
