+++ src/vi/truyenqq/src/eu/kanade/tachiyomi/extension/vi/truyenqq/TruyenQQ.kt
@@ -1,47 +1,102 @@
 package eu.kanade.tachiyomi.extension.vi.truyenqq
 
-import eu.kanade.tachiyomi.network.GET
 import eu.kanade.tachiyomi.source.model.Filter
 import eu.kanade.tachiyomi.source.model.FilterList
 import eu.kanade.tachiyomi.source.model.MangasPage
 import eu.kanade.tachiyomi.source.model.Page
 import eu.kanade.tachiyomi.source.model.SChapter
 import eu.kanade.tachiyomi.source.model.SManga
-import eu.kanade.tachiyomi.source.online.HttpSource
+import eu.kanade.tachiyomi.source.model.SMangaUpdate
 import eu.kanade.tachiyomi.util.asJsoup
 import keiyoushi.annotation.Source
+import keiyoushi.network.get
 import keiyoushi.network.rateLimit
+import keiyoushi.source.KeiSource
 import keiyoushi.utils.tryParse
+import kotlinx.serialization.json.JsonElement
 import okhttp3.CacheControl
 import okhttp3.Headers
+import okhttp3.HttpUrl
 import okhttp3.HttpUrl.Companion.toHttpUrl
 import okhttp3.OkHttpClient
-import okhttp3.Request
 import okhttp3.Response
+import org.jsoup.nodes.Document
 import java.text.SimpleDateFormat
 import java.util.Locale
 import java.util.TimeZone
 import kotlin.time.Duration.Companion.seconds
 
 @Source
-abstract class TruyenQQ : HttpSource() {
+abstract class TruyenQQ : KeiSource() {
 
-    override val supportsLatest: Boolean = true
+    override fun OkHttpClient.Builder.configureClient() = apply {
+        rateLimit(1, 2.seconds) { it.host == baseUrl.toHttpUrl().host }
+    }
 
-    override val client: OkHttpClient = network.client.newBuilder()
-        .rateLimit(1, 2.seconds) { it.host == baseUrl.toHttpUrl().host }
-        .build()
-
-    override fun headersBuilder(): Headers.Builder = super.headersBuilder().add("Referer", "$baseUrl/")
+    override fun Headers.Builder.configureHeaders(): Headers.Builder = apply {
+        add("Referer", "$baseUrl/")
+    }
 
     private val dateFormat = SimpleDateFormat("dd/MM/yyyy", Locale.ROOT).apply {
         timeZone = TimeZone.getTimeZone("Asia/Ho_Chi_Minh")
     }
 
-    // Trang html chứa popular
-    override fun popularMangaRequest(page: Int): Request = GET("$baseUrl/truyen-yeu-thich" + if (page > 1) "/trang-$page" else "", headers)
+    // ============================== Popular ===============================
 
-    override fun popularMangaParse(response: Response): MangasPage {
+    override suspend fun getPopularManga(page: Int): MangasPage {
+        val url = "$baseUrl/truyen-yeu-thich" + if (page > 1) "/trang-$page" else ""
+
+        return parseMangaPage(client.get(url, headers))
+    }
+
+    // =============================== Latest ===============================
+
+    override suspend fun getLatestUpdates(page: Int): MangasPage {
+        val url = "$baseUrl/truyen-moi-cap-nhat" + if (page > 1) "/trang-$page" else ""
+
+        return parseMangaPage(client.get(url, headers))
+    }
+
+    // =============================== Search ===============================
+
+    override suspend fun getSearchMangaList(page: Int, query: String, filters: FilterList): MangasPage {
+        val endpoint = if (query.isNotBlank()) "tim-kiem" else "tim-kiem-nang-cao"
+        val url = ("$baseUrl/$endpoint" + if (page > 1) "/trang-$page" else "").toHttpUrl().newBuilder().apply {
+            if (query.isNotBlank()) {
+                addQueryParameter("q", query)
+            } else {
+                filters.forEach { filter ->
+                    when (filter) {
+                        is CountryFilter -> addQueryParameter("country", filter.values[filter.state].id)
+                        is StatusFilter -> addQueryParameter("status", filter.values[filter.state].id)
+                        is ChapterCountFilter -> addQueryParameter("minchapter", filter.values[filter.state].id)
+                        is SortByFilter -> filter.state?.let {
+                            addQueryParameter("sort", (it.index * 2 + (if (it.ascending) 1 else 0)).toString())
+                        }
+                        is GenreList -> {
+                            addQueryParameter(
+                                "category",
+                                filter.state.asSequence()
+                                    .filter { it.state == Filter.TriState.STATE_INCLUDE }
+                                    .joinToString(",") { it.id },
+                            )
+                            addQueryParameter(
+                                "notcategory",
+                                filter.state.asSequence()
+                                    .filter { it.state == Filter.TriState.STATE_EXCLUDE }
+                                    .joinToString(",") { it.id },
+                            )
+                        }
+                        else -> {}
+                    }
+                }
+            }
+        }.build()
+
+        return parseMangaPage(client.get(url, headers))
+    }
+
+    private fun parseMangaPage(response: Response): MangasPage {
         val document = response.asJsoup()
         val manga = document.select("ul.grid > li").map { element ->
             SManga.create().apply {
@@ -55,52 +110,33 @@ abstract class TruyenQQ : HttpSource() {
         return MangasPage(manga, hasNextPage)
     }
 
-    // Trang html chứa Latest (các cập nhật mới nhất)
-    override fun latestUpdatesRequest(page: Int): Request = GET("$baseUrl/truyen-moi-cap-nhat" + if (page > 1) "/trang-$page" else "", headers)
-
-    override fun latestUpdatesParse(response: Response): MangasPage = popularMangaParse(response)
-
-    // Tìm kiếm
-    override fun searchMangaRequest(page: Int, query: String, filters: FilterList): Request {
-        val endpoint = if (query.isNotBlank()) "tim-kiem" else "tim-kiem-nang-cao"
-        val url = ("$baseUrl/$endpoint" + if (page > 1) "/trang-$page" else "").toHttpUrl().newBuilder().apply {
-            if (query.isNotBlank()) {
-                addQueryParameter("q", query)
-            } else {
-                (filters.ifEmpty { getFilterList() }).forEach { filter ->
-                    when (filter) {
-                        is CountryFilter -> addQueryParameter("country", filter.values[filter.state].id)
-                        is StatusFilter -> addQueryParameter("status", filter.values[filter.state].id)
-                        is ChapterCountFilter -> addQueryParameter("minchapter", filter.values[filter.state].id)
-                        is SortByFilter -> filter.state?.let {
-                            addQueryParameter("sort", (it.index * 2 + if (it.ascending) 1 else 0).toString())
-                        }
-                        is GenreList -> {
-                            addQueryParameter(
-                                "category",
-                                filter.state.filter { it.state == Filter.TriState.STATE_INCLUDE }
-                                    .joinToString(",") { it.id },
-                            )
-                            addQueryParameter(
-                                "notcategory",
-                                filter.state.filter { it.state == Filter.TriState.STATE_EXCLUDE }
-                                    .joinToString(",") { it.id },
-                            )
-                        }
-                        else -> {}
-                    }
-                }
+    override suspend fun getMangaByUrl(url: HttpUrl): SManga? {
+        if (url.host == baseUrl.toHttpUrl().host) {
+            client.get(url, headers).use { response ->
+                return parseMangaDetails(response.asJsoup())
             }
-        }.build()
-
-        return GET(url, headers)
+        }
+        return null
     }
 
-    override fun searchMangaParse(response: Response): MangasPage = popularMangaParse(response)
+    // =========================== Manga Details ============================
 
-    // Details
-    override fun mangaDetailsParse(response: Response): SManga = SManga.create().apply {
-        val document = response.asJsoup()
+    override suspend fun fetchMangaUpdate(
+        manga: SManga,
+        chapters: List<SChapter>,
+        fetchDetails: Boolean,
+        fetchChapters: Boolean,
+    ): SMangaUpdate {
+        client.get(getMangaUrl(manga), headers).use { response ->
+            val document = response.asJsoup()
+            val details = if (fetchDetails) parseMangaDetails(document) else manga
+            val chaptersList = if (fetchChapters) parseChapterList(document) else chapters
+
+            return SMangaUpdate(details, chaptersList)
+        }
+    }
+
+    private fun parseMangaDetails(document: Document): SManga = SManga.create().apply {
         val info = document.selectFirst(".list-info")!!
 
         title = document.select("h1").text()
@@ -124,135 +160,35 @@ abstract class TruyenQQ : HttpSource() {
         else -> SManga.UNKNOWN
     }
 
-    // Chapters
-    override fun chapterListParse(response: Response): List<SChapter> = response.asJsoup()
-        .select("div.works-chapter-list div.works-chapter-item").map { element ->
-            SChapter.create().apply {
-                setUrlWithoutDomain(element.selectFirst("a")!!.attr("href"))
-                name = element.select("a").text().trim()
-                date_upload = dateFormat.tryParse(element.select(".time-chap").text())
-            }
+    private fun parseChapterList(document: Document): List<SChapter> = document.select("div.works-chapter-list div.works-chapter-item").map { element ->
+        SChapter.create().apply {
+            setUrlWithoutDomain(element.selectFirst("a")!!.attr("href"))
+            name = element.select("a").text().trim()
+            date_upload = dateFormat.tryParse(element.select(".time-chap").text())
         }
+    }
 
-    // Pages
-    override fun pageListRequest(chapter: SChapter): Request = super.pageListRequest(chapter)
-        .newBuilder()
-        .cacheControl(CacheControl.FORCE_NETWORK)
-        .build()
+    // =============================== Pages ================================
 
-    override fun pageListParse(response: Response): List<Page> = response.asJsoup()
-        .select(".page-chapter img:not([src*='stress.gif'])")
-        .mapIndexed { idx, it ->
-            Page(idx, imageUrl = it.absUrl("src"))
+    override suspend fun getPageList(chapter: SChapter): List<Page> {
+        val cacheControl = CacheControl.FORCE_NETWORK
+        client.get(getChapterUrl(chapter), headers, cacheControl).use { response ->
+            return response.asJsoup()
+                .select(".page-chapter img:not([src*='stress.gif'])")
+                .mapIndexed { idx, it ->
+                    Page(idx, imageUrl = it.absUrl("src"))
+                }
         }
+    }
 
-    override fun imageUrlParse(response: Response): String = throw UnsupportedOperationException()
+    // ============================== Filters ===============================
 
-    // Filters
-    override fun getFilterList(): FilterList = FilterList(
+    override fun getFilterList(data: JsonElement?): FilterList = FilterList(
         Filter.Header("Không dùng chung với tìm kiếm bằng tên"),
         CountryFilter(),
         StatusFilter(),
         ChapterCountFilter(),
         SortByFilter(),
         GenreList(getGenreList()),
-    )
-
-    private class Genre(name: String, val id: String) : Filter.TriState(name) {
-        override fun toString(): String = name
-    }
-
-    private class CountryFilter :
-        Filter.Select<Genre>(
-            "Quốc gia",
-            arrayOf(
-                Genre("Tất cả", "0"),
-                Genre("Trung Quốc", "1"),
-                Genre("Việt Nam", "2"),
-                Genre("Hàn Quốc", "3"),
-                Genre("Nhật Bản", "4"),
-                Genre("Mỹ", "5"),
-            ),
-        )
-
-    private class StatusFilter :
-        Filter.Select<Genre>(
-            "Tình trạng",
-            arrayOf(
-                Genre("Tất cả", "-1"),
-                Genre("Đang tiến hành", "0"),
-                Genre("Hoàn thành", "2"),
-            ),
-        )
-
-    private class ChapterCountFilter :
-        Filter.Select<Genre>(
-            "Số lượng chương",
-            arrayOf(
-                Genre("0", "0"),
-                Genre(">= 100", "100"),
-                Genre(">= 200", "200"),
-                Genre(">= 300", "300"),
-                Genre(">= 400", "400"),
-                Genre(">= 500", "500"),
-            ),
-        )
-
-    private class SortByFilter :
-        Filter.Sort(
-            "Sắp xếp",
-            arrayOf("Ngày đăng", "Ngày cập nhật", "Lượt xem"),
-            Selection(2, false),
-        )
-
-    private class GenreList(state: List<Genre>) : Filter.Group<Genre>("Thể loại", state)
-
-    // console.log([...document.querySelectorAll(".genre-item")].map(e => `Genre("${e.innerText}", "${e.querySelector("span").dataset.id}")`).join(",\n"))
-    private fun getGenreList() = listOf(
-        Genre("Action", "26"),
-        Genre("Adventure", "27"),
-        Genre("Anime", "62"),
-        Genre("Chuyển Sinh", "91"),
-        Genre("Cổ Đại", "90"),
-        Genre("Comedy", "28"),
-        Genre("Comic", "60"),
-        Genre("Demons", "99"),
-        Genre("Detective", "100"),
-        Genre("Doujinshi", "96"),
-        Genre("Drama", "29"),
-        Genre("Fantasy", "30"),
-        Genre("Gender Bender", "45"),
-        Genre("Harem", "47"),
-        Genre("Historical", "51"),
-        Genre("Horror", "44"),
-        Genre("Huyền Huyễn", "468"),
-        Genre("Isekai", "85"),
-        Genre("Josei", "54"),
-        Genre("Mafia", "69"),
-        Genre("Magic", "58"),
-        Genre("Manhua", "35"),
-        Genre("Manhwa", "49"),
-        Genre("Martial Arts", "41"),
-        Genre("Military", "101"),
-        Genre("Mystery", "39"),
-        Genre("Ngôn Tình", "87"),
-        Genre("One shot", "95"),
-        Genre("Psychological", "40"),
-        Genre("Romance", "36"),
-        Genre("School Life", "37"),
-        Genre("Sci-fi", "43"),
-        Genre("Seinen", "42"),
-        Genre("Shoujo", "38"),
-        Genre("Shoujo Ai", "98"),
-        Genre("Shounen", "31"),
-        Genre("Shounen Ai", "86"),
-        Genre("Slice of life", "46"),
-        Genre("Sports", "57"),
-        Genre("Supernatural", "32"),
-        Genre("Tragedy", "52"),
-        Genre("Trọng Sinh", "82"),
-        Genre("Truyện Màu", "92"),
-        Genre("Webtoon", "55"),
-        Genre("Xuyên Không", "88"),
     )
 }
